Adding JWT claims

+++ resources/item.py
@@ -1,4 +1,4 @@
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
 
@@ -37,7 +37,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
